* UPD: Weather Editor Upader. Check description...
* FIX: Fixed wnd resizer for editor.
* FIX: Fixed crash when setting weather when level is not loaded.
* FIX: Fixed crash when setting time factor when level is not loaded.
* FIX: Fixed HWND values.

+++ code/editors/xrWeatherEditor/ide_impl.cpp
@@ -69,12 +69,12 @@ bool ide_impl::idle				() const
 
 HWND ide_impl::main_handle		()
 {
-	return				((HWND)m_window->Handle.ToInt32());
+	return				((HWND)m_window->Handle.ToInt64());
 }
 
 HWND ide_impl::view_handle		()
 {
-	return				((HWND)m_window->view().draw_handle().ToInt32());
+	return				((HWND)m_window->view().draw_handle().ToInt64());
 }
 
 void ide_impl::run				()
